% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVar.R
\name{$extractVar}
\alias{$extractVar}
\title{RA Method: Extract variables from an RA Object}
\description{
Method for extracting the private variables of an RA object.
}
\section{Usage}{

\preformatted{
RAobj$extractVar(nameList)
}
}

\section{Arguments}{

\describe{
\item{nameList}{A list of the variable names to be extracted from the RA object.}
}
}

\examples{
vcffile <- simDS()
rafile <- VCFtoRA(vcffile$vcf)
RAdata <- readRA(rafile)

## extract the depth matrix
depthMat <- RAdata$extractVar(list("ref","alt"))
str(depthMat)

## extract the chromosome and positions
assemblyInfo <- RAdata$extractVar(list("chrom","pos"))
str(assemblyInfo)
}
\seealso{
\code{\link{RA}}
}
\author{
Timothy P. Bilton
}
