% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSamples.R
\name{$mergeSamples}
\alias{$mergeSamples}
\title{RA Method: Merge samples in an RA object}
\description{
Method for merging read count information from multiple samples in an RA object
}
\details{
The main argument is \code{samID} which expects a character vector (length equal to the number of individuals in
the dataset) where the enteries correspond to the sample IDs. Enteries in \code{samID} that are identical
result in the corresponding samples being merged. See the examples for an illustration.
}
\section{Usage}{

\preformatted{
RAobj$mergeSamples(samID, useID=TRUE)
}
}

\section{Arguments}{

\describe{
\item{samID}{Character vector:  }
\item{useID}{Logical: If \code{TRUE}, the values in \code{samID} are used. Otherwise, the original sample
IDs in the RA object are used.}
}
}

\examples{
## Load simulated dataset with GUSbase into R
file <- simDS()
RAfile <- VCFtoRA(file$vcf)
subset <- readRA(RAfile, snpsubset = 10:30)

## create new sample ID vector
samID = substr(ra$.__enclos_env__$private$indID, 1, 2) ## sample IDs from RA object
samID[5:104] = paste0(samID[5:104],"_", 1:100)
head(samID) # will merge first row with third and second row with fourth row

## merge samples
subset$mergeSamples(samID)

}
\seealso{
\code{\link{RA}}
}
\author{
Timothy P. Bilton
}
