% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRDPlot.R
\name{$RDDPlot}
\alias{$RDDPlot}
\alias{RDDPlot}
\title{Produce a reference dosage density (RDD) plot
#'
#' Some discription
#'}
\usage{
RAobj$RDDPlot(ref, alt, ploid=2, gfreq=NULL, file=NULL, maxdepth=500, maxSNPs=1e5, ...)
}
\description{
Produce a reference dosage density (RDD) plot
#'
#' Some discription
#'
}
