% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVCF.R
\name{$writeVCF}
\alias{$writeVCF}
\title{RA Method: Write genotype data to VCF format}
\description{
Method for converting allele count data in an RA object back to VCF format.
}
\section{Usage}{

\preformatted{
RAobj$writeVCF(snpsubset=NULL, indsubset=NULL, file="GUSbase", IDuse=NULL)
}
}

\section{Arguments}{

\describe{
\item{snpsubset}{Integer vector giving the indices of the SNPs to retain in the VCF file}
\item{indsubset}{Integer vector giving the indices of the samples to retain in the VCF file}
\item{file}{Character giving the name of the VCF file to be written}
\item{IDuse}{Character vector specifying alternative samples names. Useful for anonymizing sample IDs.}
}
}

\examples{
file <- simDS()
RAfile <- VCFtoRA(file$vcf)
subset <- readRA(RAfile, snpsubset = 10:30)

## write the subset of the data back to VCF format
subset$writeVCF(file = "subset")

}
\seealso{
\code{\link{RA}}
}
\author{
Timothy P. Bilton
}
