% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cometPlot.R
\name{cometPlot}
\alias{cometPlot}
\title{#' Produce a cometPlot
#'
#' Some discription
#'
#' @usage
#' RAobj$comotPlot(model="random", alpha=NULL, filename=NULL, cex=1, maxdepth=500, ...)
#'
#' @name $cometPlot
#'
#'}
\usage{
cometPlot(ref, alt, ploid = 2, gfreq = NULL, file = NULL, cex = 1,
  maxdepth = 500, maxSNPs = 1e+05, res = 300, ...)
}
\description{
#' Produce a cometPlot
#'
#' Some discription
#'
#' @usage
#' RAobj$comotPlot(model="random", alpha=NULL, filename=NULL, cex=1, maxdepth=500, ...)
#'
#' @name $cometPlot
#'
#'
}
