% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePop.R
\name{createPop}
\alias{createPop}
\title{Create population}
\usage{
createPop(RAobj, pop = "unrelated", ploid = 2, ...)
}
\arguments{
\item{RAobj}{An object of classes 'RA' and 'R6' that was created from the \code{\link{readRA}} function}

\item{pop}{A character string specifying the type of population to be created. Currently, only \code{unrelated} and \code{full-sib}
are implemented (see details below)}

\item{ploid}{Integer specifying the ploid level in polyploids populations. Only used in \code{unrelated} populations.}

\item{...}{Additional arguments passed to \code{\link{makePop}} function.}
}
\value{
An \code{R6} object of classes \code{RA} and \code{UR} (for unrelated populations) or \code{FS} (for full-sib populations)
}
\description{
Function which creates a object for storing information related to a specific population structure.
}
\author{
Timothy P. Bilton.
}
