% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagPlot.R
\name{DiagPlots}
\alias{DiagPlots}
\alias{cometPlot}
\alias{$rocketPlot}
\alias{$cometPlot}
\alias{RRDplot}
\alias{$RRDplot}
\title{Produce a comet plot, rocket plot or RDD plot}
\usage{
\preformatted{
## Method for RA object
RAobj$cometPlot(ploid=2, filename=NULL, cex=1, maxdepth=500, maxSNPs=1e5, res=300, ...)
RAobj$rocketPlot(ploid=2, filename=NULL, cex=1, maxdepth=500, maxSNPs=1e5, res=300, scaled=TRUE, ...)
RAobj$RDDPlot(ploid=2, filename=NULL, maxdepth=500, maxSNPs=1e5, ...)
}

## Stand-alone functions
cometPlot(ref, alt, ploid=2, gfreq=NULL, file=NULL, cex=1, maxdepth=500, maxSNPs=1e5, res=300, ...)
rocketPlot(ref, alt, ploid=2, gfreq=NULL, file=NULL, cex=1, maxdepth=500, maxSNPs=1e5, res=300, scaled=TRUE, ...)
RDDPlot(ref, alt, ploid=2, gfreq=NULL, file=NULL, maxdepth=500, maxSNPs=1e5, ...)
}
\arguments{
\item{ref}{Integer matrix: Read counts for the reference allele}

\item{alt}{Integer matrix: Read counts for the alternate allele}

\item{ploid}{Non-negative integer: The ploidy level of the individuals}

\item{gfreq}{Numeric matrix: Genotype probabilities for each SNP and individual}

\item{cex}{Numeric value: A magnification value for the size of the text and labels on the axes}

\item{maxdepth}{Postive integer: The maximum depth for the x-axis and y-axis}

\item{maxSNPs}{Postive integer: The maximum number of SNPs to use in creating the graph}

\item{res}{Positive numeric value: The resolution of the graph when saving to a file}

\item{filename}{Character: The name of the file to save the plot to}

\item{scaled}{Logical: If TRUE, the counts in the rocket plot are scaled}
}
\description{
Diagnostic graphs for assessing the validity of the binomial model for read count data
generated using high-throughput sequencing technology
}
\examples{

## Read in simulated data in GUSbase package
vcffile <- simDS()$vcf
rafile <- VCFtoRA(vcffile)
radata <- readRA(rafile)

## Produce a comet plot
radata$cometPlot()

## Produce a rocket plot
radata$rocketPlot()

## RDD plot
radata$RDDPlot()

}
\author{
Timothy P. Bilton
}
