% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSbase-package.R
\docType{package}
\name{GUSbase-package}
\alias{GUSbase-package}
\alias{GUSbase}
\title{Genotyping Uncertainty with Sequencing data - Base package}
\description{
\tabular{ll}{ Package: \tab GUSbase\cr Type: \tab Package\cr Version: \tab
0.1.0\cr Date: \tab 2018-08-15\cr License: \tab GPL 3\cr }
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<tbilton@maths.otago.ac.nz>
}
\keyword{package}
