% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{logit}
\alias{logit}
\alias{inv.logit}
\alias{logit2}
\alias{inv.logit2}
\title{Transformation functions.}
\usage{
logit(p)

inv.logit(logit_p)

logit2(p)

inv.logit2(logit_p)
}
\arguments{
\item{p}{Probability value}

\item{logit_p}{Logit probability value}
}
\description{
Transformation used in GUSbase.
}
\details{
Two transformations are used in the GUSverse. These are the standard logit transformation
\deqn{logit(p) = log\left(\frac{p}{1-p}\right)}
and a modified logit transformation
\deqn{logit2(p) = log(\frac{2p}{1-2p}).}
}
