% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRA.R
\name{readRA}
\alias{readRA}
\title{Read an Reference/Alternate (RA) file.}
\usage{
readRA(rafile, snpsubset = NULL, sampthres = 0.01, excsamp = NULL, ...)
}
\arguments{
\item{rafile}{Character string giving the path to the RA file to be read into R. Typically the required string is
returned from the VCFtoRA function when the VCF file is converted to RA format.}

\item{snpsubset}{Integer vector giving the indices of the SNPs from the RA file to be read in. This indices correspond
to the rows of the RA file (excluding the header row).}

\item{sampthres}{A numeric value giving the filtering threshold for which individual samples are removed. Default is 0.01
which means that samples with an average number of reads per SNP that is less than 0.01 are removed.}

\item{excsamp}{A character vector of the sample IDs that are to be excluded (or discarded). Note that the sample IDs must correspond
to those given in the RA file that is to be processed.}

\item{...}{Additional arguments (not used).}
}
\value{
An R6 object of class RA.
}
\description{
Function which processes an RA file into an RA object.
}
\details{
RA format is a tab-delimited with columns, CHROM, POS, SAMPLES
where SAMPLES consists of sampleIDs, which typically consist of a colon-delimited sampleID, flowcellID, lane, seqlibID.
e.g.,
\tabular{llll}{
CHROM \tab  POS  \tab   999220:C4TWKACXX:7:56 \tab  999204:C4TWKACXX:7:56 \cr
1     \tab  415  \tab   5,0                   \tab  0,3                   \cr
1     \tab  443  \tab   1,0                   \tab  4,4                   \cr
1     \tab  448  \tab   0,0                   \tab  0,2
}
Note: Indels are removed, multiple alternative alleles are removed and ./. is translated into 0,0.
}
\examples{
file <- simDS()
RAfile <- VCFtoRA(file$vcf)
simdata <- readRA(RAfile)

## Reading in a subset of the data
# Takes SNPs 10 to 30
subset <- readRA(RAfile, snpsubset = 10:30)

# Read in a random set of SNPs
set.seed(675)
subset <- readRA(RAfile, snpsubset = sample(1:1000, size=10))

}
\author{
Timothy P. Bilton
}
