% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDS.R
\name{simDS}
\alias{simDS}
\title{Simulated dataset and pedigree file}
\usage{
simDS()
}
\value{
A character string of the complete path to the simulated dataset contained in the package
and a pedigree file that goes with the data.
}
\description{
Function for extracting the path to the Variant Call Format (VCF) file of a simulated dataset and associated pedigree file.
}
\details{
The data set consists of a simulated F1 family of 100 progeny, 1000 SNPs and three chromosomes. The data is in VCF format (see this \href{https://samtools.github.io/hts-specs/VCFv4.2.pdf}{page}
for specification of VCF format). The pedigree file contains five columns, namely
\itemize{
\item SampleID: A unique character string of the sample ID. These correspond to those found in the VCF file
\item IndividualID: A character giving the ID number of the individual for which the sample corresponds to.
\item Mother: The ID of the mother as given in the IndividualID. Blank fields means the mother is unknown.
\item Father: The ID of the father as given in the IndividualID. Blank fields means the father is unknown.
\item Family: The name of the Family for a group of progeny with the same parents.
}
}
\examples{
## extract the name of the vcf file and the pedigree file
simDS()
}
\author{
Timothy P. Bilton
}
