% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transformations}
\alias{transformations}
\alias{logit}
\alias{inv.logit}
\alias{logit2}
\alias{inv.logit2}
\title{Transformation functions}
\usage{
logit(p)

inv.logit(logit_p)

logit2(p)

inv.logit2(logit_p)
}
\arguments{
\item{p}{Probability value}

\item{logit_p}{Logit probability value}
}
\description{
Functions used in the GUS universe to transform parameters estimation onto an unbounded
interval. These functions are used in likelihood computations optimised using direct
maximization routines.
}
\details{
Parameter transformations used in the GUS universe:

1. The logit transformation:
\deqn{logit(p) = log(\frac{p}{1-p})}{logit(p) = log(p/(1-p))}
2. A modified logit transformation (called logit2):
\deqn{logit2(p) = log(\frac{2p}{1-2p})}{logit2(p) = log(2p/(1-2p))}
}
