% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UR.R
\name{UR}
\alias{UR}
\title{UR object}
\description{
Class for storing RA data and associated functions for analysis of unrelated populations.
}
\section{Usage}{

\preformatted{
URobj <- makeUR(RAobj, ploid = 2,indsubset = NULL, filter = list(MAF = 0.01,
  MISS = 0.5, HW = c(-0.05, Inf), MAXDEPTH = 500), mafEst = TRUE,
  nThreads = 2)
}
}

\section{Details}{

An UR object is created from the \code{\link{makeUR}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (or methods)
for analyzing the data. Information in an UR object are specific to unrelated populations (or
populations with no known relationships).
}

\seealso{
\code{\link{makeUR}}
}
\author{
Timothy P. Bilton
}
\section{Super class}{
\code{\link[GUSbase:RA]{GUSbase::RA}} -> \code{UR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UR-new}{\code{UR$new()}}
\item \href{#method-UR-clone}{\code{UR$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="RDDPlot"><a href='../../GUSbase/html/RA.html#method-RA-RDDPlot'><code>GUSbase::RA$RDDPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="cometPlot"><a href='../../GUSbase/html/RA.html#method-RA-cometPlot'><code>GUSbase::RA$cometPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="extractVar"><a href='../../GUSbase/html/RA.html#method-RA-extractVar'><code>GUSbase::RA$extractVar()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="mergeSamples"><a href='../../GUSbase/html/RA.html#method-RA-mergeSamples'><code>GUSbase::RA$mergeSamples()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="print"><a href='../../GUSbase/html/RA.html#method-RA-print'><code>GUSbase::RA$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="rocketPlot"><a href='../../GUSbase/html/RA.html#method-RA-rocketPlot'><code>GUSbase::RA$rocketPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeRA"><a href='../../GUSbase/html/RA.html#method-RA-writeRA'><code>GUSbase::RA$writeRA()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeVCF"><a href='../../GUSbase/html/RA.html#method-RA-writeVCF'><code>GUSbase::RA$writeVCF()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UR-new"></a>}}
\if{latex}{\out{\hypertarget{method-UR-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UR$new(R6obj, ploid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
