% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSbase-package.R
\docType{package}
\name{GUSbase-package}
\alias{GUSbase-package}
\alias{GUSbase}
\title{Genotyping Uncertainty with Sequencing data - Base package}
\description{
GUSbase is a package for analyzing high-throughput sequencing data via reference and alternate count information.
It forms the base to a number of other packages with specialised methods for analyzing high-throughput sequencing data:
\itemize{
\item{GUSMap: Genotyping Uncertainty with Sequencing data and linkage MAPping}
\item{GUSLD: Genotyping Uncertainty with Sequencing data - Linkage Disequilibrium}
}
}
\details{
\tabular{ll}{ Package: \tab GUSbase\cr Type: \tab Package\cr Version: \tab
0.1.1\cr Date: \tab 2018-10-29\cr License: \tab GPL 3\cr }
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<tbilton@maths.otago.ac.nz>
}
\keyword{package}
