% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUR.R
\name{makeUR}
\alias{makeUR}
\title{Make an unrelated (UR) population}
\usage{
makeUR(RAobj, indsubset = NULL, ploid = 2, filter = list(MAF = 0.01,
  MISS = 0.5, MAXDEPTH = 500), mafEst = TRUE, nClust = 2)
}
\arguments{
\item{RAobj}{Object of class RA created via the \code{\link{readRA}} function.}

\item{indsubset}{Integer vector specifying which samples of the RA dataset to retain in the UR
population.}

\item{ploid}{An integer number specifying the ploidy level of the population. Currently, only
even ploidy levels are valid.}

\item{filter}{Named list of thresholds for various criteria used to fiter SNPs.
See below for details.}

\item{mafEst}{Logical value indicating whether the allele frequences and sequencing
error parameters are to estimated for each SNP (see details).}

\item{nClust}{Integer vector specifying the number of clusters to use in the foreach loop. Only used in the estimation of
allele frequencies when \code{mafEst=TRUE}.}
}
\value{
An R6 object of class UR.
}
\description{
Create an UR object from an RA object and perform standard filtering and compute statistics
specific to unrelated populations.
}
\details{
If \code{mafEst=TRUE}, then the major allele frequency and sequencing error rate for each SNP is estimated based on optimizing the likelihood
\deqn{P(Y=a) = \sum_{G} P(Y=a|G)P(G)}
where \eqn{P(G)} are genotype probabilities under Hardy Weinberg Equilibrium (HWE) and \eqn{P(Y=a|G)} are the probilities given in Equation (5) of
\insertCite{bilton2018genetics2;textual}{GUSbase}. Otherwise, the allele
frequencies are taken as the mean of the allele ratio (defined as the number of reference reads divided by the total number of reads) and
the sequencing error rate is assumed to be zero.

The filtering criteria currently implemented are
\itemize{
\item{Minor allele frequency (MAF): }{SNPs are discarded if their MAF is less than the threshold (default is 0.01)}
\item{Proportion of missing data (MISS): }{SNPs are discarded if the proportion of individuals with no reads
(e.g. missing genotype) is greater than the threshold value (default is 0.5)}
\item{Maximum average SNP read depth (MAXDEPTH): }{SNPs are discarded if the average read depth for the SNP
is larger than the threshold (default is 500)}
}

Estimation of the allele frequencies when \code{mafEst=TRUE} is parallelized using the \code{\link[foreach]{foreach}} function, where the
number of cores to use in the parallelization is specified by the argument \code{nClust}. Note: Do not
set the number of cores to be more than what is available on your computer (or bad things will happen!).
}
\examples{
file <- simDS()
RAfile <- VCFtoRA(file$vcf)
simdata <- readRA(RAfile)

## make unrelated population
urpop <- makeUR(simdata, mafEst=FALSE)
}
\references{
\insertRef{bilton2018genetics2}{GUSbase}
}
\author{
Timothy P. Bilton
}
