% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUR.R
\name{makeUR}
\alias{makeUR}
\title{Make an unrelated (UR) population}
\usage{
makeUR(RAobj, filter = list(MAF = 0.05, MISS = 0.5), ploid = 2,
  mafEst = TRUE, nClust = 3)
}
\arguments{
\item{RAobj}{Object of class RA created via the \code{\link{readRA}} function.}

\item{filter}{Named list of of thresholds for various criteria of fitering SNPs.
See below for details.}

\item{ploid}{An integer number specifying two times the ploidy level of the population.}

\item{mafEst}{Logical value indicating whether the allele frequences and sequencing
error parameters are to estimated for each SNP.}

\item{nClust}{Integer vector specifying the number of clusters to use in the foreach loop. Only used in the estimation of
allele frequencies when \code{mafEst=TRUE}.}
}
\value{
An R6 object of class UR.
}
\description{
Create an UR object from an RA object and performs standard filtering and computes statistics specific to unrelated populations.
}
\details{
If \code{mafEst=TRUE}, then the major allele frequency and sequencing error rate for each SNP is estimated based on optimizing the likelihood
\deqn{P(Y=a) = \sum_{G} P(Y=a|G)P(G)}
where \eqn{P(G)} are genotype probabilities under Hardy Weinberg Equilibrium (HWE) and \eqn{P(Y=a|G)} are the probilities given in Equation (5) of
\insertCite{bilton2018genetics2;textual}{GUSbase}. Otherwise, the allele
frequencies are taken as the mean of the allele ratio (defined as the number of reference reads divided by the total number of reads) and
the sequencing error rate is assumed to zero.
}
\section{Filtering}{

The filtering criteria currently implemented are
\itemize{
\item{Minor allele frequency (MAF): }{SNPs are discarded if their MAF is less than the threshold (default is 0.05)}
\item{Proportion of missing data (MISS): }{SNPs are discarded if the proportion of individuals with no reads (e.g. missing genotype) is greater than the threshold value (default is 0.5)}
}
}

\references{
\insertRef{bilton2018genetics2}{GUSbase}
}
\author{
Timothy P. Bilton
}
