% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePop_UR.R
\name{makeUR}
\alias{makeUR}
\title{Make an unrelated population}
\usage{
makeUR(RAobj, filter = list(MAF = 0.05, MISS = 0.2), ploid = 1,
  mafEst = TRUE)
}
\arguments{
\item{RAobj}{Object of class RA created via the \code{\link{readRA}} function.}

\item{filter}{Named list of of thresholds for various criteria of fitering SNPs.
See below for details.}

\item{ploid}{An integer number specifying two times the ploidy level of the population.}

\item{mafEst}{Logical value indicating whether the allele frequences and sequencing
error parameters are to estimated for each SNP.}
}
\description{
Some discription
}
\details{
The filtering criteria currently implemented are
\itemize{
\item{MAF}{Minor allele frequency}
\item{MISS}{Proportion of missing data}
}
}
\author{
Timothy P. Bilton
}
