% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA.R
\docType{data}
\name{RA}
\alias{RA}
\title{RA Object}
\usage{
## Create RA object
RAobj <- readRA(genofile, gform, sampthres = 0.01, excsamp = NULL)
}
\description{
Class for storing reference/alternative (RA) data and methods for performing exporartory data analysis.
}
\section{Details}{

An RA object is returned from the \code{\link{readRA}} function and contains the RA data, various
statistics of the dataset that have been computed, and functions (or methods) for analyzing the data.
}

\seealso{
\code{\link{readRA}}
}
\author{
Timothy P. Bilton
}
\keyword{datasets}
