% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA.R
\name{RA}
\alias{RA}
\title{RA Object}
\description{
Class for storing reference/alternative (RA) data and methods for performing exporartory data analysis
on sequencing data.
}
\section{Usage}{

\preformatted{
## Create RA object
RAobj <- readRA(rafile, snpsubset=NULL, sampthres = 0.01, excsamp = NULL)

## RA Functions (Methods)
RAobj$extractVar(nameList)
RAobj$mergeSamples(samID, useID=TRUE)
RAobj$writRA(snpsubset=NULL, indsubset=NULL, file="GUSbase")
RAobj$writeVCF(snpsubset=NULL, indsubset=NULL, file="GUSbase", IDuse=NULL)
}
}

\section{Details}{

An RA object is returned from the \code{\link{readRA}} function and contains the RA data, various
statistics of the dataset that have been computed, and functions (or methods) for analyzing the data.
}

\section{Methods (Functions)}{


A list of the methods that are available to an RA object:
\describe{
    \item{\code{\link{$extractVar}}}{Extract private variables stored in an RA object}
    \item{\code{\link{$mergeSamples}}}{Merge read counts for different samples}
    \item{\code{\link{$writeRA}}}{Convert the data in the RA object back to an RA file}
    \item{\code{\link{$writeVCF}}}{Convert the data in the RA object back to VCF format}
}
}

\seealso{
\code{\link{readRA}}
}
\author{
Timothy P. Bilton
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RA$new()}}
\item \href{#method-print}{\code{RA$print()}}
\item \href{#method-mergeSamples}{\code{RA$mergeSamples()}}
\item \href{#method-extractVar}{\code{RA$extractVar()}}
\item \href{#method-cometPlot}{\code{RA$cometPlot()}}
\item \href{#method-rocketPlot}{\code{RA$rocketPlot()}}
\item \href{#method-RDDPlot}{\code{RA$RDDPlot()}}
\item \href{#method-writeVCF}{\code{RA$writeVCF()}}
\item \href{#method-writeRA}{\code{RA$writeRA()}}
\item \href{#method-clone}{\code{RA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$new(List)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeSamples"></a>}}
\if{latex}{\out{\hypertarget{method-mergeSamples}{}}}
\subsection{Method \code{mergeSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$mergeSamples(samID, useID = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extractVar"></a>}}
\if{latex}{\out{\hypertarget{method-extractVar}{}}}
\subsection{Method \code{extractVar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$extractVar(nameList)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cometPlot"></a>}}
\if{latex}{\out{\hypertarget{method-cometPlot}{}}}
\subsection{Method \code{cometPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$cometPlot(
  ploid = 2,
  filename = NULL,
  cex = 1,
  maxdepth = 500,
  maxSNPs = 1e+06,
  res = 300,
  ind = FALSE,
  ncores = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rocketPlot"></a>}}
\if{latex}{\out{\hypertarget{method-rocketPlot}{}}}
\subsection{Method \code{rocketPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$rocketPlot(
  ploid = 2,
  filename = NULL,
  cex = 1,
  maxdepth = 500,
  maxSNPs = 1e+05,
  res = 300,
  scaled = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDDPlot"></a>}}
\if{latex}{\out{\hypertarget{method-RDDPlot}{}}}
\subsection{Method \code{RDDPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$RDDPlot(ploid = 2, filename = NULL, maxdepth = 500, maxSNPs = 1e+05, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writeVCF"></a>}}
\if{latex}{\out{\hypertarget{method-writeVCF}{}}}
\subsection{Method \code{writeVCF()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$writeVCF(snpsubset = NULL, indsubset = NULL, file = "GUSbase", IDuse = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writeRA"></a>}}
\if{latex}{\out{\hypertarget{method-writeRA}{}}}
\subsection{Method \code{writeRA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$writeRA(snpsubset = NULL, indsubset = NULL, file = "GUSbase")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
