% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA.R
\docType{data}
\name{RA}
\alias{RA}
\title{RA Object}
\description{
Class for storing reference/alternative (RA) data and methods for performing exporartory data analysis
on sequencing data.
}
\section{Usage}{

\preformatted{
## Create RA object
RAobj <- readRA(rafile, sampthres = 0.01, excsamp = NULL)

## RA Functions (Methods)
RAobj$extractVar(nameList)
RAobj$writeVCF(snpsubset=NULL, indsubset=NULL, file="GUSbase", IDuse=NULL)
}
}

\section{Details}{

An RA object is returned from the \code{\link{readRA}} function and contains the RA data, various
statistics of the dataset that have been computed, and functions (or methods) for analyzing the data.
}

\section{Methods (Functions)}{


A list of the methods that are available to an RA object:
\describe{
    \item{\code{\link{$extractVar}}}{Extract private variables stored in an RA object}
    \item{\code{\link{$writeVCF}}}{Convert the data in the RA object back to VCF format}
}
}

\seealso{
\code{\link{readRA}}
}
\author{
Timothy P. Bilton
}
\keyword{datasets}
