% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UR.R
\docType{data}
\name{UR}
\alias{UR}
\title{UR object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
URobj <- makeUR()
}
\description{
Class for unrelated populations
}
\details{
An UR object is created from the \code{\link{makeUR}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (or methods)
for analyzing the data. Information in an UR are specific to unrelated populations.
}
\seealso{
\code{\link{makeUR}}
}
\author{
Timothy P. Bilton
}
\keyword{datasets}
